+++ src/router/routes/users/handlers/put.ts
@@ -1,10 +1,10 @@
 import { IncomingMessage } from "http";
 import { validate } from 'uuid';
 
-import { MESSAGE_INVALID_ID, MESSAGE_WRONG_USER_DATA } from "../../../constants/index.ts";
-import UserStore from '../../../store/index.ts';
-import { TServerResponse } from "../../../types/index.ts";
-import { parseReqParams, sendData } from "../../../utils/index.ts";
+import { MESSAGE_INVALID_ID, MESSAGE_WRONG_USER_DATA } from "../../../../constants/index.ts";
+import UserStore from '../../../../store/index.ts';
+import { TServerResponse } from "../../../../types/index.ts";
+import { parseReqParams, sendData } from "../../../../utils/index.ts";
 
 export const handlePut = (
   req: IncomingMessage,
